後はCSS整える

+++ resources/views/posts/index.blade.php
@@ -4,28 +4,33 @@
    
 
 <!-- 送信先と方法の指定 -->
+<div id="tweetform">
 <form action="/tweet" method="post">
 @csrf
-<!-- 入力方法、初期表示する内容placeholderの指定、画像の表示 -->
+<!-- 入力方法、初期表示する内容placeholderの指定、画像の表示 -->  
     <input type="text" name="tweet" placeholder="何をつぶやこうか...?">
     <input type="image" src="/images/post.png" alt="送信する">
 </form>
+</div>
+<div id="tweets">
 <table>
 @foreach($lists as $list)
 <tr>
     <td>
-        <img src="/images/{{ $list->user->images }}" alt="icon" class="circle2">
-        <!-- 投稿の表示 -->
+        <img src="/images/{{ $list->user->images }}" alt="icon" class="circle2">       
     </td>
-    <br>
+    <!-- 投稿の表示 -->
     <td>{{ $list->user->username }}</td>
     <td>{{ $list->posts }}</td>
     <td>{{ $list->created_at }}</td>
-    
-    <td><a href="" data-toggle="modal" data-target="#modal{{$list->id}}"><img src="/images/edit.png" alt="edit" ></a></td>
-    <td><a href="/posts/{{$list->id}}/delete" onclick="return confirm('このつぶやきを削除します。よろしいでしょうか？')"><img src="/images/trash.png" alt="trash"></a></td> 
-</tr>
 
+    <!-- ログイン中のユーザー分のみ編集・削除可能 (=== 値とデータの型が一緒なら表示) -->
+@if($list->user_id === Auth::id())   
+    <td><a href="" data-toggle="modal" data-target="#modal{{$list->id}}"><img src="/images/edit.png" alt="edit" ></a></td>
+    <td><a href="/posts/{{$list->id}}/delete" onclick="return confirm('このつぶやきを削除します。よろしいでしょうか？')"><img src="/images/trash.png" alt="trash"></a></td>
+@endif
+</tr>
+</div>
 
 <!-- 送信したツイートの受け取り -->
